openocd: Add recipe to support flash tools
Richard,

This is the first pass at openocd, it builds both the native and target
versions, but I don't have a jtag device to test it with.

There was one change I needed to make might be another funky bug, in libftdi
I needed to change as follows:

-DEPENDS_virtclass-native = "virtual/libusb0-native"
+DEPENDS_virtclass-native = "virtual/libusb-native0"

It seems the 0 moved around somehow, I have not looked at it yet.

Sau!

Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>