Add shared make jobserver support
This is a WIP to add a make job server into bitbake. This means the pool
of make tokens is central and shared by all tasks executed instead of the
current one job pool per task. Currently we can end up with many more
make subprocesses executing that is intended or optimal.

Implementation wise, make usually uses a pipe for this functionality. Here
we substitute a named pipe (fifo) and intercept the make commands, passing
in file descriptors to the central fifo. This assumes knowledge of make's
internal API, on the plus side it hasn't changed since 1999.

Looking to the future we could dynamically control the pool but one step
at a time.

TODO:
  * Remove hardcoded /tmp/makefifo and use something in TMPDIR or similar
    (alongside the lock file?)
  * Remove hardcoded make threads number and set from PARALLEL_MAKE
  * If PARALLEL_MAKE = "", don't set MAKEARGS (currently parallelism is
    set everywhere) (need to check for -j in make commandline)

I'm sending this out so at least the code is available to people. Its
not ready for merging in its current form but might be the basis for
someone else to finish this up, it also gives us something we can test
the performance implications with.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>