lib/bb/build: add linecount progress handler
Many tasks do not produce progress output, but a lot produce some kind of
output throughout their execution; if we know the number of lines to
expect and then measure the number of lines currently written out
against it we have a rough idea of the progress of the task. Add a
linecount progress handler that can be used to measure this
("linecount:<total number of lines>").

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>