qemu.inc: enable libusb support
When doing kernel development we often use USB-to-I2C/SPI/GPIO bridges
and we need to connect them to the QEMU machine to test our driver. This
feature is already supported by QEMU but, in order for this to work,
libusb has to be enabled first.

With this change, one can export an USB device to QEMU machine by
appending '-usbdevice host:vendor_id:product_id' to qemuparams.

For example:

$ runqemu qemux86 core-image-minimal kvm nographic qemuparams="-usbdevice host:vendor_id:product_id"

'lsusb' command will give you the "vendor_id:product_id" of the USB
device to be exported.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>