buildstats: Drop get_bn/set_pn and just use BUILDNAME
As far as I can tell these functions add complexity for no good reason.
There are some completely unused lines like e.getPkgs() and BUILDNAME
should not change through out the build so we can just query it.

If BUILDNAME did change, that would be a bug and it should be fixed
elsewhere.

Also take the opportunity to share some common code since the function
now has the eventmask.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>