toaster: add cancel button to UI
Add cancel button to UI. This button appears when the
build is in progress. So when this button gets clicked
cancelBuild function gets called with appropriate
arguments. The cancel button would be visible only
when the bitbake task executes.

The cancel button will be invisible for Command line
builds. The reason for that is explained in #9005.

[YOCTO #6787]

Signed-off-by: Sujith H <sujith.h@gmail.com>