patchtest: allow local mboxes and commits
This is a huge patch, not atomic at all that accomplishes several things,
with the goal of allowing mboxes and local commits to be used by patchtest.
This is an important change, making the tool less dependent on patchwork on
the input part. The test fixture located on PatchTestBase is now moved to
patchtest, so it will only be executed once, instead of every time a test
case inherited from PatchTestBase. The repo class have been refactored, with
the purpose of having an simpler and smaller API. Several command lines
introduced, including --commit, --branch, --mbox and --keep-branch. Command
line examples can be found on the github site [1]

[1] https://github.com/lsandoval/patchtest/issues/1

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>