toaster: add project field to CustomImagePackage
BuildInfoHelper was attempting to add packages to custom images,
supplying a project field so that the package only appears for
custom images inside the supplied project.

This field was missing from the CustomImagePackage model, which
was causing errors to be thrown.

Add the field to the model, and the migration to add it the the
database. Ensure that when CustomImagePackages are looked up,
the queryset is constrained to the project for the current build.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>