toaster: adjust build dashboard for failed builds
Remove the side bar and build details modules for failed builds.

Modify templates so that the span is moved out of the individual
elements and into the container. This makes it simpler to change
the width of the main container, depending on whether the
build failed or not (which in turn hides or shows the left-hand
menu).

[YOCTO #8443]

Signed-off-by: Elliot Smith <elliot.smith@intel.com>