toaster-manual: remove all references to 'build mode'
Soon after the 2.0 release Toaster modes were removed. This patch
updates the latest Toaster manual to remove all references to 'modes',
in preparation for the 2.1 release version of the manual.

The changes involved rewriting section 1.1, and small changes to sections

4.1.1.2.2
4.3
4.3.1
4.4.4

This patch also removes references to the third toasterconf.json file,
which is no longer used to set up production instances of Toaster.

Signed-off-by: Belen Barros Pena <belen.barros.pena@intel.com>