buildhistory.bbclass: Break up the find to catch failures
If for some reason the 'find' operation were to fail, the buildhistory would
not fail, but instead store invalid history!

[YOCTO #9031]

Note this fixes the symptom, but not the cause of 9031.  That is a separate
issue, see the defect for more details on the work being done to resolve the
cause of the failure.

(From OE-Core rev: 0c3efaabec2223185f2697a60ae72addffc2ec2d)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>