repo: add MboxItem class and medium refactor
MboxItem class was added to represent file or url based mbox items
listed inside Repo.items. It shouldn't be directly instantiated,
as there are subclasses (MboxFileItem and MboxURLItem) which
implement the necessary methods.

Internally items are stored in Repo._mboxitems and then the items
@property method has been modified to return this variable.

The refactor is not yet complete, but it has progressed.

Signed-off-by: Alex Franco <alejandro.franco@linux.intel.com>