toaster: ui fixes for build cancellation
Move 'cancel' button to the rigt of the % of tasks completed for builds
in progress. Also, make sure that, once the build starts, we show the
'cancel' button for any builds started from the web UI, and an
explanation of why the 'cancel' button is not there for builds started
from the command line.

[YOCTO #6787]

Signed-off-by: Belen Barros Pena <belen.barros.pena@intel.com>
Signed-off-by: Sujith H <sujith.h@gmail.com>