toaster: add data-field attributes to edit columns dropdown
ToasterTable IDs are generated on the fly depending on the data being
rendered. When the "builds" and "projects" tables were converted to
ToasterTable, the IDs changed as they were being generated by
ToasterTable; this broke the Selenium test suite, which was
querying for page elements by ID.

Add data-field attributes to the <li> elements in the
edit columns drop-down for ToasterTables (e.g. data-field="started_on").
This is preferable to using IDs, as an ID can only occur on a single
element, and the location of an ID may occasionally change to support
Toaster's JavaScript APIs.

By contrast, data-* attributes can be added with little cost, and there
is a good chance that they will be maintained over a longer period
in the HTML.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>