toaster: "all builds" is no longer sortable by recipe or time
During conversion to ToasterTable, sorting by time was not
implemented for "all builds", as it had been disabled in the old
Toaster templates.

Sorting by recipe name was disabled, as this was not trivial to
port to ToasterTable (that will require modifications to the
ToasterTable API and/or how we retrieve data for the table).

However, the Selenium tests test for sorting by time and recipe
name.

Remove these columns from the test.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>