toaster: use data-field attributes to find checkboxes in edit columns
Rather than use the IDs of the checkboxes in the edit columns
drop-down, use the data-field attributes on the <li> elements
to locate the checkbox.

Also make a more generic methods which gets the text content
of a set of elements selected by a CSS selector and use this
to construct the lists under test.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>