bitbake.conf: add findutils-native to ASSUME_PROVIDED
It's possible for findutils-native to get built.  There's no point in this as
this is part of the expected host platform but this can introduce races or even
bugs (4.5.19 appears to have a leaking fd bug, resulting in asserts) so add it
to ASSUME_PROVIDED so it definitely won't get built.

(From OE-Core rev: b753dae334641480cb4a232ce240f9f56be5568f)

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>