toaster: fix command-line builds sticking at 100%
Command-line builds were completing correctly but not being stored
by Toaster. This was because a conditional in the update_build_object()
method assumed that all builds would have a buildrequest property,
causing a failure for command-line builds.

Fix the conditional to ensure that a build has a buildrequest property
before testing the value of that property.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>