patchtest: repo: simplify setup/clean methods and split pass/fail status of merges
Another big change, with several changes included. The original idea was to split
pass/fail series merges, but I started to do changes here and there, specially on
the sub-fuctions related to setup and clean. Also, the argument that indicates to
ignore merge failures was removed: the testing will be done no matter if there
are merge fails.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>