native.bbclass: poison TARGET_CFLAGS etc for native builds
TARGET_* FLAGS are never passed to the native toolchain. However,
they are exported so their values affect the checksum of every task.

Replace with fixed values to both avoid any unnecessary rebuilding
of native packages when target flags are changed (as before) and now
also to poison any accidental use of target flags with the native
toolchain.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>