uninative: correctly enable uninative
The previous attempt at soft-failing when uninative was enabled didn't actually
work, because the workers didn't evaluate the function that actually enabled
uninative.

In a BuildStarted handler we can check if we need to download or extract the
uninative tarball.

In a ConfigParsed handler on the workers we can check if the uninative loader is
present, and if so enable it.

(From OE-Core rev: 6e7b3c9889dbb77cd75106c124d9b01aabc8027b)

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>