toaster-tests: migrate to Selenium for UI tests
Create a new folder for Selenium tests.

Add a new base Selenium testcase class and a helper which
instantiates a webdriver for a given browser.

Add a sample Selenium test case which can be used as a template
for creating new tests.

Amend the pip requirements file to include Selenium.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>