toaster: use empty token
If client xmlrpc token is not provided in the command line, bitbake
generates random token. Server token in --server-only mode is always
empty. This prevents clients with non-empty tokens to communicate
with the server. Specifying empty token should prevent generation of
random token and makes it possible to communicate with the server.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>