matchbox-keyboard: Hide desktop launcher
Add patch that hides the keyboard desktop launcher, remove patch that
tries and fails to make the keyboard a single-instance application.

The desktop launcher of matchbox-keyboard is a source of far more
problems than solutions: As an example there's supposed to be only
one instance running at a time but we give the user several ways to
start multiple instances (and the Matchbox WM Single-Instance
implementation is broken by both design and implementation).

After this patch the only instance of matchbox-keyboard is the
daemonized one that can be shown/hidden with the panel applet (when
there is not hardware keyboard). If an additional matchbox-keyboard
needs to be started for debug reasons, it can still be done from
command line.

Fixes [YOCTO #3093].

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>